+++ src/client/App.jsx
@@ -19,6 +19,7 @@ import {
 } from '@mui/material/styles'
 import { ColorModeProvider } from './contexts/ColorModeContext.jsx';
 import { LoadingProvider } from './contexts/LoadingContext.jsx';
+import { GlobalStateProvider } from './contexts/GlobalStateContext.jsx';
 
 export const App = () => {
     // check color mode preference
@@ -49,12 +50,14 @@ export const App = () => {
 
     // render jsx
     return(
-        <LoadingProvider>
-            <ColorModeProvider value={colorMode}>
-                <ThemeProvider theme={theme}>
-                    <div>agenda-display</div>
-                </ThemeProvider>
-            </ColorModeProvider>
-        </LoadingProvider>
+        <GlobalStateProvider>
+            <LoadingProvider>
+                <ColorModeProvider value={colorMode}>
+                    <ThemeProvider theme={theme}>
+                        <div>agenda-display</div>
+                    </ThemeProvider>
+                </ColorModeProvider>
+            </LoadingProvider>
+        </GlobalStateProvider>
     );
 }